move localization to basic test
