--- AudioCuesheetEditor.End2EndTests/Tests/Smartphone/LocalizationTestSmartphone.cs
@@ -1,24 +1,0 @@
-﻿using AudioCuesheetEditor.End2EndTests.Models;
-using Microsoft.Playwright;
-
-namespace AudioCuesheetEditor.End2EndTests.Tests.Smartphone
-{
-    [TestClass]
-    public class LocalizationTestSmartphone : PlaywrightTestBase
-    {
-        protected override string? DeviceName => "iPhone 13";
-
-        [TestMethod]
-        public async Task ChangeLanguage_ShouldShowGermanHeadings_WhenGermanIsSelected()
-        {
-            var bar = new AppBar(TestPage);
-            await bar.GotoAsync();
-            await bar.ChangeLanguageAsync("German (Germany)");
-            await Expect(TestPage.GetByRole(AriaRole.Heading, new() { Name = "Abschnitte" })).ToBeVisibleAsync();
-            await Expect(TestPage.GetByRole(AriaRole.Heading, new() { Name = "Allgemeine Informationen" })).ToBeVisibleAsync();
-            await Expect(TestPage.GetByText("Aufnahmeansicht")).ToBeVisibleAsync();
-            await Expect(TestPage.GetByRole(AriaRole.Heading, new() { Name = "Titel" })).ToBeVisibleAsync();
-            await Expect(TestPage.GetByRole(AriaRole.Heading, new() { Name = "Wiedergabe" })).ToBeVisibleAsync();
-        }
-    }
-}
